refactor: renamed all service and pages to file consistence.
